function limit()
- modified function to cope with double and int

+++ src/utils/utils.h
@@ -1,7 +1,7 @@
 #ifndef UTILS_H
 #define UTILS_H
 
-extern int limit (int const x);
+extern double limit (double const x);
 extern int get_output_filename (char* outfile, char* const path, char* const prefix, char* const ext, int const attr);
 
 #endif // UTILS_H
